feat: add playing state and merge timer score function to main
