+++ index.html
@@ -128,7 +128,6 @@
       </symbol>
     </svg>
 
-    <script type="module" src="./js/timer.js"></script>
     <script type="module" src="/main.js"></script>
   </body>
 </html>
