--- js/score.js
@@ -1,5 +1,0 @@
-const scoreElem = document.querySelector('#score');
-
-export default function updateScore(score) {
-  scoreElem.textContent = 'SCORE: ' + score;
-}
